refactor: view를 반환하는 api를 한 컨트롤러로 통합
