+++ src/main/java/cart/controller/ItemViewController.java
@@ -10,13 +10,20 @@ import org.springframework.web.bind.annotation.GetMapping;
 import java.util.List;
 
 @Controller
-public class AdminController {
+public class ItemViewController {
 
     private final ItemService itemService;
 
     @Autowired
-    public AdminController(ItemService itemService) {
+    public ItemViewController(ItemService itemService) {
         this.itemService = itemService;
+    }
+
+    @GetMapping("/")
+    public String showItemList(final Model model) {
+        List<ItemResponse> items = itemService.loadAllItem();
+        model.addAttribute("products", items);
+        return "index";
     }
 
     @GetMapping("/admin")
